Move Output model into models.py
It was getting too large to live in views.py
