+++ sacro/utils.py
@@ -1,0 +1,9 @@
+from urllib.parse import urlencode
+
+from django.urls import reverse
+
+
+def reverse_with_params(param_dict, *args, **kwargs):
+    """Wrapper for django reverse that adds query parameters"""
+    url = reverse(*args, **kwargs)
+    return url + "?" + urlencode(param_dict)
